feat: useQuery 위치 변경
