+++ package.json
@@ -17,6 +17,7 @@
     "jwt-encode": "^1.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-infinite-scroller": "^1.2.6",
     "react-modal": "^3.16.1",
     "react-query": "^3.39.3",
     "react-router-dom": "^6.21.2",
@@ -52,6 +53,7 @@
     ]
   },
   "devDependencies": {
+    "@types/react-infinite-scroller": "^1.2.5",
     "prettier": "^3.2.1"
   }
 }
